pyproject.toml
[tool.poetry]
name = "railway-fastapi-template"
version = "0.1.0"
description = "This example starts up a FastAPI server"
authors = ["DeviousLab <deviouslab@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.110.2"
uvicorn = "0.30.1"
paddleocr = "2.7.0.3"
paddlepaddle = "2.6.1"
python-dotenv = "1.0.1"
httpx = "0.27.0"
python-multipart = "0.0.9"

[tool.poetry.dev-dependencies]
pytest = "8.3.3"
pytest-asyncio = "0.23.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
